updated few things
