+++ src/components/Footer.tsx
@@ -22,29 +22,29 @@ const Footer = () => {
           {/* Quick Links */}
           <div className="space-y-6">
             <h3 className="font-poppins font-bold text-xl mb-6">Quick Links</h3>
-            <ul className="space-y-3">
+            <ul className="space-y-2">
               <li>
-                <Link href="/academics" className="font-open-sans text-white/80 hover:text-accent transition-colors duration-300">
+                <Link href="/academics" className="font-open-sans text-sm text-white/80 hover:text-accent transition-colors duration-300">
                   Academics
                 </Link>
               </li>
               <li>
-                <Link href="/campus-life" className="font-open-sans text-white/80 hover:text-accent transition-colors duration-300">
+                <Link href="/campus-life" className="font-open-sans text-sm text-white/80 hover:text-accent transition-colors duration-300">
                   Campus Life
                 </Link>
               </li>
               <li>
-                <Link href="/clubs" className="font-open-sans text-white/80 hover:text-accent transition-colors duration-300">
+                <Link href="/clubs" className="font-open-sans text-sm text-white/80 hover:text-accent transition-colors duration-300">
                   Clubs
                 </Link>
               </li>
               <li>
-                <Link href="/gallery" className="font-open-sans text-white/80 hover:text-accent transition-colors duration-300">
+                <Link href="/gallery" className="font-open-sans text-sm text-white/80 hover:text-accent transition-colors duration-300">
                   Gallery
                 </Link>
               </li>
               <li>
-                <Link href="/holidays" className="font-open-sans text-white/80 hover:text-accent transition-colors duration-300">Holidays</Link>
+                <Link href="/holidays" className="font-open-sans text-sm text-white/80 hover:text-accent transition-colors duration-300">Holidays</Link>
               </li>
             </ul>
           </div>
@@ -52,24 +52,24 @@ const Footer = () => {
           {/* Policies */}
           <div className="space-y-6">
             <h3 className="font-poppins font-bold text-xl mb-6">Policies</h3>
-            <ul className="space-y-3">
+            <ul className="space-y-2">
               <li>
-                <Link href="https://drive.google.com/file/d/1GJRLD1LiBNI0-kN9IHBKNRToOCdYjtNw/view?usp=sharing" className="font-open-sans text-white/80 hover:text-accent transition-colors duration-300">
+                <Link href="https://drive.google.com/file/d/1GJRLD1LiBNI0-kN9IHBKNRToOCdYjtNw/view?usp=sharing" className="font-open-sans text-sm text-white/80 hover:text-accent transition-colors duration-300">
                   Code of Conduct
                 </Link>
               </li>
               <li>
-                <Link href="https://drive.google.com/file/d/1lv3-vPzostvvHCLLhooXRIl4pd5lEXfw/view?usp=sharing" className="font-open-sans text-white/80 hover:text-accent transition-colors duration-300">
+                <Link href="https://drive.google.com/file/d/1lv3-vPzostvvHCLLhooXRIl4pd5lEXfw/view?usp=sharing" className="font-open-sans text-sm text-white/80 hover:text-accent transition-colors duration-300">
                   Attendance Policy
                 </Link>
               </li>
               <li>
-                <Link href="https://drive.google.com/file/d/1yZnQDkKua2D1oWQBtHU_xRq8auXrKtGu/view?usp=sharing" className="font-open-sans text-white/80 hover:text-accent transition-colors duration-300">
+                <Link href="https://drive.google.com/file/d/1yZnQDkKua2D1oWQBtHU_xRq8auXrKtGu/view?usp=sharing" className="font-open-sans text-sm text-white/80 hover:text-accent transition-colors duration-300">
                   Scholarship Policy
                 </Link>
               </li>
               <li>
-                <Link href="https://drive.google.com/file/d/1wxkUKD_MFo22ZPpxEDy8CgKdTuHN16wP/view?usp=sharing" className="font-open-sans text-white/80 hover:text-accent transition-colors duration-300">
+                <Link href="https://drive.google.com/file/d/1wxkUKD_MFo22ZPpxEDy8CgKdTuHN16wP/view?usp=sharing" className="font-open-sans text-sm text-white/80 hover:text-accent transition-colors duration-300">
                   Examination Policy
                 </Link>
               </li>
@@ -79,34 +79,34 @@ const Footer = () => {
           {/* Other Policies */}
           <div className="space-y-6">
             <h3 className="font-poppins font-bold text-xl mb-6">Other Policies</h3>
-            <ul className="space-y-3">
+            <ul className="space-y-2">
               <li>
-                <Link href="https://drive.google.com/file/d/1TN_2XOzlgvgFMcewR_LnYdMRfuY0KGOv/view?usp=sharing" className="font-open-sans text-white/80 hover:text-accent transition-colors duration-300">
+                <Link href="https://drive.google.com/file/d/1TN_2XOzlgvgFMcewR_LnYdMRfuY0KGOv/view?usp=sharing" className="font-open-sans text-sm text-white/80 hover:text-accent transition-colors duration-300">
                   Library Policy
                 </Link>
               </li>
               <li>
-                <Link href="https://drive.google.com/file/d/14XRHr6I1h8oHipR9tbdXJdahoN21Almb/view?usp=sharing" className="font-open-sans text-white/80 hover:text-accent transition-colors duration-300">
+                <Link href="https://drive.google.com/file/d/14XRHr6I1h8oHipR9tbdXJdahoN21Almb/view?usp=sharing" className="font-open-sans text-sm text-white/80 hover:text-accent transition-colors duration-300">
                   Klib Rules
                 </Link>
               </li>
               <li>
-                <Link href="https://drive.google.com/file/d/1rwl0LJtZlq9W_LsKuJ6tmKDbA8YyiHle/view?usp=sharing" className="font-open-sans text-white/80 hover:text-accent transition-colors duration-300">
+                <Link href="https://drive.google.com/file/d/1rwl0LJtZlq9W_LsKuJ6tmKDbA8YyiHle/view?usp=sharing" className="font-open-sans text-sm text-white/80 hover:text-accent transition-colors duration-300">
                   Sexual Harassment Policy
                 </Link>
               </li>
               <li>
-                <Link href="https://drive.google.com/file/d/1K8Z4nVVk6_Py8ja_uC3gmYEtNP9owuYM/view?usp=sharing" className="font-open-sans text-white/80 hover:text-accent transition-colors duration-300">
+                <Link href="https://drive.google.com/file/d/1K8Z4nVVk6_Py8ja_uC3gmYEtNP9owuYM/view?usp=sharing" className="font-open-sans text-sm text-white/80 hover:text-accent transition-colors duration-300">
                   Anti-Ragging Policy
                 </Link>
               </li>
               <li>
-                <Link href="https://drive.google.com/file/d/1dSdkRn0HPCpyjITGtBvvrYjrNe1icvEj/view?usp=sharing" className="font-open-sans text-white/80 hover:text-accent transition-colors duration-300">
+                <Link href="https://drive.google.com/file/d/1dSdkRn0HPCpyjITGtBvvrYjrNe1icvEj/view?usp=sharing" className="font-open-sans text-sm text-white/80 hover:text-accent transition-colors duration-300">
                   Social Media Policy
                 </Link>
               </li>
               <li>
-                <Link href="https://drive.google.com/file/d/1mf76AtfWKcxtYJJIjedBTXR5XHreT49L/view?usp=sharing" className="font-open-sans text-white/80 hover:text-accent transition-colors duration-300">
+                <Link href="https://drive.google.com/file/d/1mf76AtfWKcxtYJJIjedBTXR5XHreT49L/view?usp=sharing" className="font-open-sans text-sm text-white/80 hover:text-accent transition-colors duration-300">
                   Career Services Policy
                 </Link>
               </li>
@@ -116,7 +116,7 @@ const Footer = () => {
           {/* Contact Info */}
           <div className="space-y-6">
             <h3 className="font-poppins font-bold text-xl mb-6">Contact</h3>
-            <div className="space-y-3">
+            <div className="space-y-2">
               <p className="font-open-sans text-white/80">
                 PWIOI Campus, Bengaluru
               </p>
